[revshare] support storing multiple revshare configs in one plugin (#56)
Add a key parameter to revshare operations to support multiple configs for one community. For example a community might have a rev share config for the "dev_core_team" and a separate one for the "content_moderators" that are used differently on the revenue-generating platform. This change is backwards-compatible with last version of revshare plugin.